[FEAT] Skip question if it keeps getting filtered
The script will now skip a question if it can't say it without
activating the filter. Previously it would keep on saying the options,
even if the players had no idea what the question was about.

Also updated alt wording for correct answer. "Answer was x" to
"[Player name filtered] ...". Players were mistakingly thinking that
the time had run out with the prevoius wording.