[FIX] Fix autoplay bug & mode selection
Set the default allowed categories for autoplay to
 every category instead of an empty table. This
fixes the bug where if you turned on autoplay
without selecting any categories the script would
get stuck in an infinite loop and get a stack
overflow error.

Also fix a bug where selecting the quiz mode
from the dropdown menu was ineffective. The mode
variable was being set to the options table
instead of options[1].