[FEAT] Add a version checking system
Add a system that gets the latest version from the GitHub repo, and
sends out a notification if the latest version doesn't match the
VERSION variable. The notification gives an option to copy the latest
loadstring version of the script and a "Don't show again" button.

Add this because people often copy the source code instead of using
loadstring to get the script from GitHub, which results in them not
getting any future updates.