[FEAT] Category list no longer gets filtered
The category list no longer gets filtered when being sent in chat. This
is because the version that gets sent by default is now a simple list
which only sends the names of available categories instead of
distinguishing between different quizzes and difficulties. There is a
toggle to send the old detailed category list in chat; however, it is
not recommended to turn it on as it will almost certainly cause filter
issues.

Quizzes under the same difficulty and category no longer get a suffix
with their difficulty name if only one difficulty exists for the given
category. Only index gets appended. This applies both to the UI
dropdown and detailed category list. For instance, instead of:
history-easy1, history-easy2, it is now simply: history-1, history-2.

This change has made me add a difficulty to every quiz, as it will no
longer clutter the lists with unnecessary suffixes.

Difficulties that include 0 quizzes no longer get displayed in the
autoplay filter dropdown.

Complete rework of 'English' quiz. Addition of brainrot and brainrot2.