[FIX] Infinite autoplay loop if no filter selected
Fix autoplay looping infinitely (or until a stack overflow error is
triggered). This happened if a difficulty was present in the
difficultyOrder table but did not include any quizzes.

Fix typo in the leaderboard tab (I don't know how I didn't notice this
earlier).

Add a notificaton if the writeToDataFile() functon is called but
writefile is not suppored by the executor. Credit to IY for the idea.