[FEAT] 6 new quizzes, bug fixes, and UI changes
Add six new quizzes.

Fix the filtering of the timeout message not yielding, which caused the
bot to send the next question before the timeout message was resent.

Remove unused pointManger ClearQuizPointsForPlayer and
ClearGlobalPointsForPlayer functions.

UI leaderboard labels are now always updated when points get modified
or reset and when players leave.

Fix single quizzes with empty difficulties getting a "-1" suffix.

Fix the leaderboard labels missing a space after the medal emoji when
initially created.

Target label is now set to "Target: None" when the search doesn't find
a target.

Remove ability to select current quiz points when manually adding
points. The feature was confusing and not useful.