[FEAT] Make custom categories case insensitive
When a custom category is being created, it is now lowercased and
compared to every existing category. If a category with the same name
is found, the new quiz gets added to the existing category instead of
creating a separate category. This prevents two categories with the
same name but different casing from being created.

Tables with strings are now sorted alphabetically using a custom
function that lowers the case of the strings before comparing them.
Previously, the strings were only sorted by their ASCII values, which
would always put strings that started with a lowercase letter at the
end.