[CHG] Fully migrate to TextChatService
Fully migrate to the new TextChatService, as Roblox now forces it on
all games. The old code using
ReplicatedStorage.DefaultChatSystemChatEvents no longer works in any
games, even those which have textChatService.ChatVersion set to
"LegacyChatService". This fixes the issue with messages not sending on
games which previously used the deprecated LegacyChatService.

Wrap built-in quiz definitions in do-end block to prevent reaching the
200 local variable limit. This is necessary as more quizzes are added,
especially by users who modify the source code directly instead of
using CustomCategoryManger. Thank you to lortiz.frfr for reporting this.

Update Rayfield stable fork to latest version.

Use Rayfield's "ShowText" setting to display "Show quizbot" instead of
the default "Show Rayfield" in the button which appears when Rayfield
is hidden.